requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Pillow==10.0.1
tkintermapview==1.29
